refactor : board API 구조 변경
