+++ src/main/java/com/zerobase/school/board/dto/CreateBoardRequest.java
@@ -1,3 +1,3 @@
-package com.zerobase.school.board;
+package com.zerobase.school.board.dto;
 public class CreateBoardRequest {
     private Long bId;}
