--- src/main/java/com/zerobase/school/board/readme.md
@@ -1,12 +1,0 @@
-### <출처>:
-[1] 제로베이스 Part7, Part11, 2023.11.13->전반적
-
-[2] 제로베이스, Part14. Q11-15, Q41-45, Q56-60 답지, 2023.11.13->전반적
-
-[3] 제로베이스, Part14. Q95-100 답지, 2023.11.20->전반적
-
-[4] 출처: 계란, 계란 블로그,“Account7 Ch8”, https://oranthy.tistory.com/m/270, 2023.10.10->전반적
-
-[5] 출처: 계란, 계란 블로그,”Account7 Ch7”, https://oranthy.tistory.com/m/269, 2023.10.10->전반적
-
-[6] 제로베이스, hi.anna, 어제 오늘 내일, "현재 날짜, 현재 시간 구하기", 2021.06.18, https://hianna.tistory.com/607
