+++ src/main/java/com/zerobase/school/board/repository/BoardRepository.java
@@ -1,4 +1,5 @@
-package com.zerobase.school.board;
+package com.zerobase.school.board.repository;
+import com.zerobase.school.board.domain.Board;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
